Configuration restructure.
