+++ conf/identifiers.go
@@ -1,0 +1,22 @@
+package conf
+
+const (
+	planIdentifier = "plan"
+	keepIdentifier = "keep"
+	pathIdentifier = "path"
+)
+
+const (
+	keepFor    = "for"
+	keepLatest = "latest"
+)
+
+const (
+	blockStart = "{"
+	blockEnd   = "}"
+)
+
+const (
+	commentBashStyle = '#'
+	commentCStyle    = "//"
+)
